tree JSON to State function
